unknown

+++ Distraction_Tracker/Form1.cs
@@ -129,7 +129,7 @@ namespace Distraction_Tracker
 
             // Calculate average attention duration from List and convert back into Timespan
             Double averageDuration = this.distractionDurations.Count > 0 ? this.distractionDurations.Average() : 0.0;
-            TimeSpan averageTimespanDuration = new TimeSpan(Convert.ToInt32(averageDuration));
+            TimeSpan averageTimespanDuration = new TimeSpan(Convert.ToInt64(averageDuration));
 
             // Set AverageAttention label text property
             AverageAttention.Text = averageTimespanDuration.ToString("hh\\:mm\\:ss");
